go implementation w/ RSS
