--- .github/workflows/actions.yml
@@ -1,36 +1,0 @@
-name: Aggregate Smartraveller Data
-
-on:
-  # schedule:
-  #  - cron: '0 */6 * * *'
-  workflow_dispatch: {}
-
-jobs:
-  aggregate:
-    name: Aggregate Smartraveller Destinations and Advisories
-    runs-on: ubuntu-latest
-    permissions:
-      # Give GITHUB_TOKEN write permission to commit and push changes
-      contents: write
-    steps:
-      - name: Checkout repository
-        uses: actions/checkout@v4
-
-      - name: Set up Python 3.12
-        uses: actions/setup-python@v2
-        with:
-          allow-prereleases: true
-          python-version: 3.12
-
-      - name: Install requirements
-        run: |
-          pip install pipenv
-          pipenv install --deploy
-
-      - name: Run aggregation script
-        run: pipenv run python ./smartraveller/aggregate.py
-
-      - name: Commit aggregated data
-        uses: stefanzweifel/git-auto-commit-action@v5
-        with:
-          commit_message: "ci: update Smartraveller data"
