+++ .vscode/launch.json
@@ -2,11 +2,11 @@
     "version": "0.2.0",
     "configurations": [
         {
-            "name": "Start Smartraveller API",
+            "name": "Start API Server",
             "type": "go",
             "request": "launch",
             "mode": "debug",
-            "program": "${workspaceFolder}/api/api.go",
+            "program": "${workspaceFolder}/main.go",
         }
     ]
 }
